Move utils to single shared file
